[REFACTOR] BaseTimeEntity 추가 및 entity에 적용
