+++ src/main/java/com/smeme/server/model/BaseTimeEntity.java
@@ -1,0 +1,26 @@
+package com.smeme.server.model;
+
+
+import com.fasterxml.jackson.annotation.JsonFormat;
+import jakarta.persistence.EntityListeners;
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.annotation.LastModifiedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
+
+@MappedSuperclass
+@EntityListeners(AuditingEntityListener.class)
+@Getter
+public abstract class BaseTimeEntity {
+
+    @JsonFormat(timezone = "Asia/Seoul")
+    @CreatedDate
+    protected LocalDateTime createdAt;
+
+    @JsonFormat(timezone = "Asia/Seoul")
+    @LastModifiedDate
+    protected LocalDateTime updatedAt;
+}
