+++ src/main/java/com/smeme/server/model/Member.java
@@ -2,6 +2,7 @@ package com.smeme.server.model;
 
 import static jakarta.persistence.GenerationType.*;
 
+import com.smeme.server.model.badge.Badge;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
@@ -64,6 +65,10 @@ public class Member {
         this.socialId = socialId;
         this.targetLang = targetLang;
         this.goal = null;
+    }
+
+    public void addWelcomeBadge(MemberBadge welcomeBadge) {
+        this.badges.add(welcomeBadge);
     }
 
     public void updateRefreshToken(String refreshToken) {
