+++ src/main/java/com/smeme/server/repository/badge/BadgeRepository.java
@@ -1,7 +1,8 @@
-package com.smeme.server.repository;
+package com.smeme.server.repository.badge;
 
 import com.smeme.server.model.badge.Badge;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BadgeRepository extends JpaRepository<Badge, Long> {
+
 }
