+++ src/main/java/com/smeme/server/util/message/ErrorMessage.java
@@ -36,7 +36,11 @@ public enum ErrorMessage {
 	// Training Time
 	EMPTY_TRAINING_TIME("존재하지 않는 학습계획 입니다."),
 	// badge
-	INVALID_BADGE("유효하지 않은 뱃지입니다.")
+	INVALID_BADGE("유효하지 않은 뱃지입니다."),
+	EMPTY_BADGE("존재하지 않는 뱃지입니다."),
+
+	// goal
+	EMPTY_GOAL("존재하지 않는 목표입니다.")
 	;
 
 	private final String message;
